修复倒序seek时unsigned翻转问题

+++ core/src/sublogview.h
@@ -6,8 +6,8 @@ class SubLog;
 
 struct SubLogPos
 {
-    size_t blockIndex;
-    size_t lineIndex;
+    int blockIndex;
+    int lineIndex;
 };
 
 class SubLogView : public LogView {
